Change size_t to unsigned in one place.
git-svn-id: http://svn.boost.org/svn/boost/trunk@22868 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/boost/program_options/detail/value_semantic.hpp
@@ -132,7 +132,7 @@ namespace boost { namespace program_options {
         }
         std::vector<T>* tv = boost::any_cast< std::vector<T> >(&v);
         assert(tv);
-        for (size_t i = 0; i < s.size(); ++i)
+        for (unsigned i = 0; i < s.size(); ++i)
         {
             try {
                 tv->push_back(boost::lexical_cast<T>(s[i]));
